Remove `/src` from import paths
This was added to assist VSCode in finding typings for each module
without needing to rebuild dependencies. As it turns out, Code is smart
enough to pull typings from the source without the need to rebuild.

By updating these paths we can run each module as 100% JavaScript when
built -- TypeScript is just used to compile, not to run.